Move redirect-generation out into its own source file
