+++ src/redirectParams.js
@@ -1,0 +1,28 @@
+import _ from 'lodash';
+
+function redirectParams(record, resources) {
+  const obj = {};
+
+  if (record.source === 'kb') {
+    obj._path = `/eholdings/titles/${record.id}`;
+    obj.searchType = 'titles';
+    obj.q = _.get(resources, ['query', 'query']);
+    obj.query = null;
+    // TODO: reseting filters is a temp solution until
+    // https://issues.folio.org/browse/UISE-67 is fixed
+    obj.filters = '';
+  } else if (record.source === 'local') {
+    obj._path = `/inventory/view/${record.id}`;
+    obj.searchType = null;
+    obj.q = null;
+    // The 'query' parameter should already be correct
+    // TODO: reseting filters is a temp solution until
+    // https://issues.folio.org/browse/UISE-67 is fixed
+    obj.filters = '';
+  } else {
+    return null;
+  }
+  return obj;
+}
+
+export default redirectParams;
